Moved command handling out of TelnetServer.py
When a command is received via a telnet connection, the logic behind
parsing the instruction has now been moved to a separate class in the
new file `TelnetServerOutputController.py`. This has been done for the
following reasons:

- In the future there could be many more commands which would take up
alot of space, so the specifically server-related code should be
isolated from that
- With the new class, more focus can be put on each of the commands;
Instead of separating commands with a long if-elif-else statement, each
command now has its own function with a decorator to label it as a
command
- A dedicated class to commands also makes certain tasks easier, such as
the help command which can now just look at the list of commands and
dynamically read their doc-strings to give descriptions to the user.
- This also opens up opportunities for sub-commands; With the previous
if-elif-else logic, there was little space for nested commands and
certainly no chance of documenting them with the help command. This new
setup is a step closer in this direction, perhaps with more complex
decorators to handle nested commands; Perhaps having the nested commands
separated into new subclasses that act as the command function, maybe by
utilising the __call__ function? This is just speculation at the moment,
but this is a step closer in the right direction.