Added 'modules list' command to TelnetServer
This command gives a list of the currently running modules (At the
moment there is no way to stop/pause modules, however that is a
feature to come in the near future)

The Main.getmodules() function has also been updated to now keep track
of any folders in the modules/ directory which were deemed invalid. It
saves them in a new list, self.bad_modules, in 2-length-tuples of the
directory name and a brief reason for why they were deemed invalid.

This can be useful for debugging if a module is not being loaded.

The setup of the function is a work-in-progress, as I am unhappy with
how the outputcontroller handles the 'list' subcommand of 'modules'.
This will be amended with better subcommand logic in future commits.