Changed module thread structure
Previous method:
- The main part of this would consist of the queue variable in the main
  class. Whenever a modules was to be ran, it would simply be added to
  the end of the queue.
- The ModuleRunner class held a certain number of threads (Defined by
  a class variable) which would run constantly, always looking for items
  to take off the queue and run.
- When the server shuts down, ModuleRunner sends a None object to the
  queue for every worker thread and when they each receive one, they
  close themselves to allow the program to end.

Problems with this setup / Areas for improvement:
- The program will always have the worker threads running, even when
there are no tasks to complete. This means that memory and performance
are being wasted by useless loops.
- Even when tasks need to be completed, three separate threads is in
most cases overkill besides in certain situations when many different
modules fall on the same time interval.
- The threads run both when looking for modules and when running them,
when in reality the looking for modules is already pretty much handled
in the mainloop thread. These worker threads are supposed to only be
used for the running of the modules to allow these independant modules
to now prevent the rest of the program from running smoothly.

Current Method:
- Instead of adding modules to a queue when they need to be ran, modules
are ran directly. However the difference is that now the Module class
itself handles individual threads which open to run the task and then
end once it is finished. This means that the only worker threads open are
those that are running a module.
- Instead of having a set number of worker threads running at a time,
this setup has a max number of threads defined by a new replacement
class variable in ModuleRunner: `MAX_THREADS`. Before running a module,
it will check how many threads are currently running and if has reached
the max, the module will be added to the backlog.
- The backlog is a queue object in the ModuleRunner. A separate thread
runs (only when the backlog is not empty) and waits until the number of
running threads goes below the max, only then to start the next module
in the queue. The backlog thread starts as soon as something is added to
the queue and closes once it is empty.

Implementations / Changes to go:
- All of the `modules` commands from the TelnetServer had to be changed,
but most only required a few minor reference updates. The `modules run`
command however has been commented out for now as it requires more
changes to fix.
- The threads need better error handling abilities. This should now be
easier to implement as most of it should be able to be handled in the
Module class.