WIP: Server records telnet connections
A new block has been added to the program: The Stats object.
This object should have variable references to all areas of the program
and act as a hub of information about the server. A branch of the object
is available to provide formatted strings with visuals such as tables.
This is specifically for the telnet server to make it easier to send
back server condition information.

This commit is slightly unfinished but still a working program. The few
things left to complete are:
- Remove the duplicate functions (format_timedelta & format_columns)
  which have now been moved to static functions of the StatsDecorator
  class.
- Redo the format_columns function into a more general table formatter
  with column headers and easier syntax.
- Redo branching of decorated properties (At the moment they are simply
  prefixed with a 'd_'). My standing at the moment is to move them into a
  separate object or something else.
- Separate stat classes (TelnetConnectionRecord, Stats, StatsRecorder,
  StatsDecorator & StatsController) into different files.